refactor: :truck: rename index to bugsnag
