+++ packages/node/src/bugsnag.ts
